Personal Trainer/requirements.txt
python-dotenv
openai 